Consolidate Clip-related code - Move common Clip-related code from Graphics.Unix and Graphics.Windows to Graphics - Move P/Invoke methods to GdiplusNative
